+++ src/Rule/AbstractRule.php
@@ -33,6 +33,38 @@ abstract class AbstractRule implements RuleInterface
     protected $error = '';
 
     /**
+     * List of default message patterns in case of error.
+     * supported default keys are %code%, %key% and %value%
+     *
+     * @var array
+     */
+    protected $messages = [
+        self::UNDEFINDED_CODE => 'Code message %code% is undefined',
+        self::EMPTY_KEY => '%key% is required and should not be empty: %value%',
+        self::INVALID_BIC_LIMIT => '%key%: %value% has an invalid length',
+        self::INVALID_BIC_UPPER => '%key%: %value% should be uppercase',
+        self::INVALID_BIC_ALNUM => '%key%: %value% should be alphanumeric',
+        self::INVALID_BIC_BC => '%key%: %value% has an invalid bank code',
+        self::INVALID_BIC_CC => '%key%: %value% has an invalid country code',
+        self::INVALID_BOOL => '%key%: %value% is not a valid boolean',
+        self::INVALID_DATE => '%key%: %value% is not a valid date',
+        self::INVALID_DATE_FORMAT => '%key%: %value% does not have a valid format: %format% or separator: %separator%',
+        self::INVALID_COMPARE => '%key%: %value% is not %label% %expected%',
+        self::INVALID_EMAIL => '%key%: %value% is not a valid email',
+        self::INVALID_IP => '%key%: %value% is not a valid IP',
+        self::INVALID_IP_FLAG => 'Filter IP flag: %flag% is not valid',
+        self::INVALID_JSON => '%key%: %value% is not a valid json format',
+        self::INVALID_PATTERN => '%key%: %value% does not match %pattern%',
+        self::INVALID_NUMERIC => '%key%: %value% is not numeric',
+        self::INVALID_NUMERIC_LT => '%key%: %value% is less than %min%',
+        self::INVALID_NUMERIC_GT => '%key%: %value% is greater than %max%',
+        self::INVALID_RANGE => '%key%: %value% is out of range %range%',
+        self::INVALID_STRING => '%key% does not have a string value: %value%',
+        self::INVALID_STRING_LENGTH => '%key%: The length of %value% is not between %min% and %max%',
+        self::INVALID_TIME => '%key%: %value% is not a valid time',
+    ];
+
+    /**
      * If key's context is required or not.
      * If required, value should not be empty.
      * Default sets to false.
@@ -50,6 +82,11 @@ abstract class AbstractRule implements RuleInterface
 
         if (isset($params['trim'])) {
             $this->trim = (bool) $params['trim'];
+        }
+
+        if (isset($params['messages'])) {
+            // replace existant by given messages
+            $this->messages = array_merge($this->messages, $params['messages']);
         }
 
         $this->setValue($value);
@@ -81,8 +118,7 @@ abstract class AbstractRule implements RuleInterface
 
         if ($this->isRequired()) {
             // Value is empty and required.
-            $this->error = "{$this->key} is required and should not be empty";
-            return RuleInterface::ERROR;
+            return $this->setAndReturnError(self::EMPTY_KEY);
         }
 
         // Value is empty but not required.
@@ -123,5 +159,37 @@ abstract class AbstractRule implements RuleInterface
         return $this->value === null ||
         $this->value === [] ||
         $this->value === '';
+    }
+
+    /**
+     * @return int RuleInterface::ERROR
+     */
+    protected function setAndReturnError(string $errorCode, array $replace = []): int
+    {
+        $message = $this->messages[$errorCode] ?? $this->messages[self::UNDEFINDED_CODE];
+
+        // + is used to add unexistant keys
+        $replace += [
+            '%key%'   => $this->key,
+            '%value%' => $this->canonize($this->value),
+            '%code%'  => $errorCode,
+        ];
+
+        $this->error = str_replace(array_keys($replace), $replace, $message);
+
+        return RuleInterface::ERROR;
+    }
+
+    protected function canonize($value): string
+    {
+        if (is_object($value) && ! in_array('__toString', get_class_methods($value))) {
+            return get_class($value) . ' object';
+        }
+
+        if (is_array($value)) {
+            return str_replace(["\n", "\r"], '', var_export($value, true));
+        }
+
+        return (string) $value;
     }
 }