+++ tests/Rule/DateRuleTest.php
@@ -55,27 +55,39 @@ class DateRuleTest extends TestCase
 
         yield 'Given value 29/02/2017 should not be valid' => [
             '29/02/2017', ['format' => 'dd/mm/yyyy'], RuleInterface::ERROR,
-            'date: 29/02/2017 is not valid. Check your date, format and separator'
+            "date: 29/02/2017 is not a valid date"
         ];
 
         yield 'Given value 0/02/2017 should not be valid' => [
             '0/02/2017', ['format' => 'dd/mm/yyyy'], RuleInterface::ERROR,
-            'date: 0/02/2017 is not valid. Check your date, format and separator'
+            "date: 0/02/2017 is not a valid date"
         ];
 
         yield 'Given value 0/02 should not be valid' => [
             '0/02', ['format' => 'dd/mm/yyyy'], RuleInterface::ERROR,
-            'date: 0/02 is not valid. Check your date, format and separator'
+            "date: 0/02 is not a valid date"
+        ];
+
+        yield 'Given value separator . should not be valid' => [
+            '12/02/2017', ['separator' => '.'], RuleInterface::ERROR,
+            "date: 12/02/2017 does not have a valid format: array (  0 => 'dd/mm/yyyy',) or separator: ."
         ];
 
         yield 'Given value format dd/dd/mm should not be valid' => [
             '12/02/2017', ['format' => 'dd/dd/mm'], RuleInterface::ERROR,
-            'date: 12/02/2017 is not valid. Check your date, format and separator'
+            "date: 12/02/2017 does not have a valid format: array (  0 => 'dd/dd/mm',) or separator: [,-./]"
         ];
 
         yield 'Given value format dd/ss/mm should not be valid' => [
             '12/02/2017', ['format' => 'dd/ss/mm'], RuleInterface::ERROR,
-            'date: 12/02/2017 is not valid. Check your date, format and separator'
+            "date: 12/02/2017 does not have a valid format: array (  0 => 'dd/ss/mm',) or separator: [,-./]"
+        ];
+
+        yield 'Given value format dd/ss/mm should not be valid, with specific message' => [
+            '12/02/2017', ['format' => 'dd/ss/mm', 'messages' => [
+                RuleInterface::INVALID_DATE_FORMAT => '%key% has invalid format: %format%'
+            ]],
+            RuleInterface::ERROR, "date has invalid format: array (  0 => 'dd/ss/mm',)"
         ];
     }
 }